refactor(ContextMenu): Break up into partials
