+++ packages/react-component-library/src/components/ContextMenu/partials/StyledIcon.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+import { selectors } from '@royalnavy/design-tokens'
+
+const { spacing, color } = selectors
+
+export const StyledIcon = styled.div`
+  display: inline-flex;
+  align-items: center;
+  margin-right: ${spacing('2')};
+
+  svg {
+    color: ${color('neutral', '300')};
+  }
+`
